+++ install.ps1
@@ -1,4 +1,4 @@
-$version="0.21.1"
+$version="0.22.0"
 
 if ([Environment]::Is64BitProcess) {
   $binary_arch="amd64"