+++ src/constants.go
@@ -10,7 +10,7 @@ import (
 
 const (
 	// Current version
-	version = "0.21.1"
+	version = "0.22.0"
 
 	// Core
 	coordinatorDelayMax  time.Duration = 100 * time.Millisecond